split PE resource
