+++ src/main/java/de/tum/cit/aet/artemis/programming/web/ProgrammingExerciseCreationUpdateResource.java
@@ -1,0 +1,405 @@
+package de.tum.cit.aet.artemis.programming.web;
+
+import static de.tum.cit.aet.artemis.core.config.Constants.PROFILE_CORE;
+import static de.tum.cit.aet.artemis.core.config.Constants.PROFILE_THEIA;
+
+import java.io.IOException;
+import java.net.URI;
+import java.net.URISyntaxException;
+import java.nio.file.Path;
+import java.util.ArrayList;
+import java.util.Objects;
+import java.util.Optional;
+
+import org.eclipse.jgit.api.errors.GitAPIException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Lazy;
+import org.springframework.context.annotation.Profile;
+import org.springframework.core.env.Environment;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PatchMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+
+import de.tum.cit.aet.artemis.athena.api.AthenaApi;
+import de.tum.cit.aet.artemis.core.domain.Course;
+import de.tum.cit.aet.artemis.core.domain.User;
+import de.tum.cit.aet.artemis.core.exception.BadRequestAlertException;
+import de.tum.cit.aet.artemis.core.exception.ConflictException;
+import de.tum.cit.aet.artemis.core.exception.ContinuousIntegrationException;
+import de.tum.cit.aet.artemis.core.exception.EntityNotFoundException;
+import de.tum.cit.aet.artemis.core.repository.UserRepository;
+import de.tum.cit.aet.artemis.core.security.Role;
+import de.tum.cit.aet.artemis.core.security.annotations.EnforceAtLeastEditor;
+import de.tum.cit.aet.artemis.core.service.AuthorizationCheckService;
+import de.tum.cit.aet.artemis.core.service.course.CourseService;
+import de.tum.cit.aet.artemis.core.service.feature.Feature;
+import de.tum.cit.aet.artemis.core.service.feature.FeatureToggle;
+import de.tum.cit.aet.artemis.core.util.HeaderUtil;
+import de.tum.cit.aet.artemis.exercise.service.ExerciseService;
+import de.tum.cit.aet.artemis.lecture.api.SlideApi;
+import de.tum.cit.aet.artemis.programming.domain.ProgrammingExercise;
+import de.tum.cit.aet.artemis.programming.repository.ProgrammingExerciseRepository;
+import de.tum.cit.aet.artemis.programming.service.AuxiliaryRepositoryService;
+import de.tum.cit.aet.artemis.programming.service.ProgrammingExerciseCreationUpdateService;
+import de.tum.cit.aet.artemis.programming.service.ProgrammingExerciseRepositoryService;
+import de.tum.cit.aet.artemis.programming.service.ProgrammingExerciseTaskService;
+import de.tum.cit.aet.artemis.programming.service.ProgrammingExerciseValidationService;
+import de.tum.cit.aet.artemis.programming.service.StaticCodeAnalysisService;
+import io.jsonwebtoken.lang.Arrays;
+
+/**
+ * REST controller for managing ProgrammingExercise.
+ */
+@Profile(PROFILE_CORE)
+@Lazy
+@RestController
+@RequestMapping("api/programming/")
+public class ProgrammingExerciseCreationUpdateResource {
+
+    private static final Logger log = LoggerFactory.getLogger(ProgrammingExerciseCreationUpdateResource.class);
+
+    private static final String ENTITY_NAME = "programmingExercise";
+
+    private final Environment environment;
+
+    @Value("${jhipster.clientApp.name}")
+    private String applicationName;
+
+    private final CourseService courseService;
+
+    private final AuthorizationCheckService authCheckService;
+
+    private final ExerciseService exerciseService;
+
+    private final ProgrammingExerciseValidationService programmingExerciseValidationService;
+
+    private final ProgrammingExerciseCreationUpdateService programmingExerciseCreationUpdateService;
+
+    private final ProgrammingExerciseRepositoryService programmingExerciseRepositoryService;
+
+    private final ProgrammingExerciseTaskService programmingExerciseTaskService;
+
+    private final StaticCodeAnalysisService staticCodeAnalysisService;
+
+    private final AuxiliaryRepositoryService auxiliaryRepositoryService;
+
+    private final Optional<AthenaApi> athenaApi;
+
+    private final Optional<SlideApi> slideApi;
+
+    private final ProgrammingExerciseRepository programmingExerciseRepository;
+
+    private final UserRepository userRepository;
+
+    public ProgrammingExerciseCreationUpdateResource(ProgrammingExerciseRepository programmingExerciseRepository, UserRepository userRepository,
+            AuthorizationCheckService authCheckService, CourseService courseService, ExerciseService exerciseService,
+            ProgrammingExerciseValidationService programmingExerciseValidationService, ProgrammingExerciseCreationUpdateService programmingExerciseCreationUpdateService,
+            ProgrammingExerciseRepositoryService programmingExerciseRepositoryService, ProgrammingExerciseTaskService programmingExerciseTaskService,
+            StaticCodeAnalysisService staticCodeAnalysisService, AuxiliaryRepositoryService auxiliaryRepositoryService, Optional<AthenaApi> athenaApi, Environment environment,
+            Optional<SlideApi> slideApi) {
+        this.programmingExerciseValidationService = programmingExerciseValidationService;
+        this.programmingExerciseCreationUpdateService = programmingExerciseCreationUpdateService;
+        this.programmingExerciseTaskService = programmingExerciseTaskService;
+        this.programmingExerciseRepository = programmingExerciseRepository;
+        this.userRepository = userRepository;
+        this.courseService = courseService;
+        this.authCheckService = authCheckService;
+        this.exerciseService = exerciseService;
+        this.programmingExerciseRepositoryService = programmingExerciseRepositoryService;
+        this.staticCodeAnalysisService = staticCodeAnalysisService;
+        this.auxiliaryRepositoryService = auxiliaryRepositoryService;
+        this.athenaApi = athenaApi;
+        this.environment = environment;
+        this.slideApi = slideApi;
+    }
+
+    /**
+     * POST /programming-exercises/setup : Set up a new programmingExercise (with all needed repositories etc.)
+     *
+     * @param programmingExercise the programmingExercise to set up
+     * @return the ResponseEntity with status 201 (Created) and with body the new programmingExercise, or with status 400 (Bad Request) if the parameters are invalid
+     */
+    @PostMapping("programming-exercises/setup")
+    @EnforceAtLeastEditor
+    @FeatureToggle(Feature.ProgrammingExercises)
+    public ResponseEntity<ProgrammingExercise> createProgrammingExercise(@RequestBody ProgrammingExercise programmingExercise) {
+        log.debug("REST request to setup ProgrammingExercise : {}", programmingExercise);
+
+        // Valid exercises have set either a course or an exerciseGroup
+        programmingExercise.checkCourseAndExerciseGroupExclusivity(ENTITY_NAME);
+        Course course = courseService.retrieveCourseOverExerciseGroupOrCourseId(programmingExercise);
+        authCheckService.checkHasAtLeastRoleInCourseElseThrow(Role.EDITOR, course, null);
+        programmingExerciseValidationService.validateNewProgrammingExerciseSettings(programmingExercise, course);
+
+        // Check that only allowed athena modules are used
+        athenaApi.ifPresentOrElse(api -> api.checkHasAccessToAthenaModule(programmingExercise, course, ENTITY_NAME), () -> programmingExercise.setFeedbackSuggestionModule(null));
+
+        try {
+            // Setup all repositories etc
+            ProgrammingExercise newProgrammingExercise = programmingExerciseCreationUpdateService.createProgrammingExercise(programmingExercise);
+
+            // Create default static code analysis categories
+            if (Boolean.TRUE.equals(programmingExercise.isStaticCodeAnalysisEnabled())) {
+                staticCodeAnalysisService.createDefaultCategories(newProgrammingExercise);
+            }
+
+            return ResponseEntity.created(new URI("/api/programming/programming-exercises/" + newProgrammingExercise.getId())).body(newProgrammingExercise);
+        }
+        catch (IOException | URISyntaxException | GitAPIException | ContinuousIntegrationException e) {
+            log.error("Error while setting up programming exercise", e);
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
+                    .headers(HeaderUtil.createAlert(applicationName, "An error occurred while setting up the exercise: " + e.getMessage(), "errorProgrammingExercise")).body(null);
+        }
+    }
+
+    /**
+     * PUT /programming-exercises : Updates an existing updatedProgrammingExercise.
+     *
+     * @param updatedProgrammingExercise the programmingExercise that has been updated on the client
+     * @param notificationText           to notify the student group about the update on the programming exercise
+     * @return the ResponseEntity with status 200 (OK) and with body the updated ProgrammingExercise, or with status 400 (Bad Request) if the updated ProgrammingExercise
+     *         is not valid, or with status 500 (Internal Server Error) if the updated ProgrammingExercise couldn't be saved to the database
+     */
+    @PutMapping("programming-exercises")
+    @EnforceAtLeastEditor
+    @FeatureToggle(Feature.ProgrammingExercises)
+    public ResponseEntity<ProgrammingExercise> updateProgrammingExercise(@RequestBody ProgrammingExercise updatedProgrammingExercise,
+            @RequestParam(value = "notificationText", required = false) String notificationText) throws JsonProcessingException {
+        log.debug("REST request to update ProgrammingExercise : {}", updatedProgrammingExercise);
+        if (updatedProgrammingExercise.getId() == null) {
+            throw new BadRequestAlertException("Programming exercise cannot have an empty id when updating", ENTITY_NAME, "noProgrammingExerciseId");
+        }
+
+        updatedProgrammingExercise.validateGeneralSettings();
+
+        // Valid exercises have set either a course or an exerciseGroup
+        updatedProgrammingExercise.checkCourseAndExerciseGroupExclusivity(ENTITY_NAME);
+        programmingExerciseValidationService.validateStaticCodeAnalysisSettings(updatedProgrammingExercise);
+
+        // fetch course from database to make sure client didn't change groups
+        var user = userRepository.getUserWithGroupsAndAuthorities();
+        Course course = courseService.retrieveCourseOverExerciseGroupOrCourseId(updatedProgrammingExercise);
+        authCheckService.checkHasAtLeastRoleInCourseElseThrow(Role.EDITOR, course, user);
+
+        programmingExerciseValidationService.checkProgrammingExerciseForError(updatedProgrammingExercise);
+
+        var programmingExerciseBeforeUpdate = programmingExerciseRepository.findForUpdateByIdElseThrow(updatedProgrammingExercise.getId());
+        if (!Objects.equals(programmingExerciseBeforeUpdate.getShortName(), updatedProgrammingExercise.getShortName())) {
+            throw new BadRequestAlertException("The programming exercise short name cannot be changed", ENTITY_NAME, "shortNameCannotChange");
+        }
+        if (!Objects.equals(programmingExerciseBeforeUpdate.isStaticCodeAnalysisEnabled(), updatedProgrammingExercise.isStaticCodeAnalysisEnabled())) {
+            throw new BadRequestAlertException("Static code analysis enabled flag must not be changed", ENTITY_NAME, "staticCodeAnalysisCannotChange");
+        }
+        // Check if theia Profile is enabled
+        if (Arrays.asList(this.environment.getActiveProfiles()).contains(PROFILE_THEIA)) {
+            // Require 1 / 3 participation modes to be enabled
+            if (!Boolean.TRUE.equals(updatedProgrammingExercise.isAllowOnlineEditor()) && !Boolean.TRUE.equals(updatedProgrammingExercise.isAllowOfflineIde())
+                    && !updatedProgrammingExercise.isAllowOnlineIde()) {
+                throw new BadRequestAlertException("You need to allow at least one participation mode, the online editor, the offline IDE, or the online IDE", ENTITY_NAME,
+                        "noParticipationModeAllowed");
+            }
+        }
+        else {
+            // Require 1 / 2 participation modes to be enabled
+            if (!Boolean.TRUE.equals(updatedProgrammingExercise.isAllowOnlineEditor()) && !Boolean.TRUE.equals(updatedProgrammingExercise.isAllowOfflineIde())) {
+                throw new BadRequestAlertException("You need to allow at least one participation mode, the online editor or the offline IDE", ENTITY_NAME,
+                        "noParticipationModeAllowed");
+            }
+        }
+
+        // Verify that the checkout directories have not been changed. This is required since the buildScript and result paths are determined during the creation of the exercise.
+        programmingExerciseValidationService.validateCheckoutDirectoriesUnchanged(programmingExerciseBeforeUpdate, updatedProgrammingExercise);
+
+        // Verify that the programming language supports the selected network access option
+        programmingExerciseValidationService.validateDockerFlags(updatedProgrammingExercise);
+
+        // Verify that a theia image is provided when the online IDE is enabled
+        if (updatedProgrammingExercise.isAllowOnlineIde() && updatedProgrammingExercise.getBuildConfig().getTheiaImage() == null) {
+            throw new BadRequestAlertException("You need to provide a Theia image when the online IDE is enabled", ENTITY_NAME, "noTheiaImageProvided");
+        }
+        // Forbid changing the course the exercise belongs to.
+        if (!Objects.equals(programmingExerciseBeforeUpdate.getCourseViaExerciseGroupOrCourseMember().getId(),
+                updatedProgrammingExercise.getCourseViaExerciseGroupOrCourseMember().getId())) {
+            throw new ConflictException("Exercise course id does not match the stored course id", ENTITY_NAME, "cannotChangeCourseId");
+        }
+        // Forbid conversion between normal course exercise and exam exercise
+        exerciseService.checkForConversionBetweenExamAndCourseExercise(updatedProgrammingExercise, programmingExerciseBeforeUpdate, ENTITY_NAME);
+
+        // Check that only allowed Athena modules are used
+        athenaApi.ifPresentOrElse(api -> api.checkHasAccessToAthenaModule(updatedProgrammingExercise, course, ENTITY_NAME),
+                () -> updatedProgrammingExercise.setFeedbackSuggestionModule(null));
+        // Changing Athena module after the due date has passed is not allowed
+        athenaApi.ifPresent(api -> api.checkValidAthenaModuleChange(programmingExerciseBeforeUpdate, updatedProgrammingExercise, ENTITY_NAME));
+
+        // Ignore changes to the default branch
+        updatedProgrammingExercise.getBuildConfig().setBranch(programmingExerciseBeforeUpdate.getBuildConfig().getBranch());
+
+        if (updatedProgrammingExercise.getAuxiliaryRepositories() == null) {
+            // make sure the default value is set properly
+            updatedProgrammingExercise.setAuxiliaryRepositories(new ArrayList<>());
+        }
+
+        // Update the auxiliary repositories in the DB and ProgrammingExercise instance
+        auxiliaryRepositoryService.handleAuxiliaryRepositoriesWhenUpdatingExercises(programmingExerciseBeforeUpdate, updatedProgrammingExercise);
+
+        // Update the auxiliary repositories in the VCS. This needs to be decoupled to break circular dependencies.
+        programmingExerciseRepositoryService.handleAuxiliaryRepositoriesWhenUpdatingExercises(programmingExerciseBeforeUpdate, updatedProgrammingExercise);
+
+        if (updatedProgrammingExercise.getBonusPoints() == null) {
+            // make sure the default value is set properly
+            updatedProgrammingExercise.setBonusPoints(0.0);
+        }
+
+        // Only save after checking for errors
+        ProgrammingExercise savedProgrammingExercise = programmingExerciseCreationUpdateService.updateProgrammingExercise(programmingExerciseBeforeUpdate,
+                updatedProgrammingExercise, notificationText);
+
+        exerciseService.logUpdate(updatedProgrammingExercise, updatedProgrammingExercise.getCourseViaExerciseGroupOrCourseMember(), user);
+        exerciseService.updatePointsInRelatedParticipantScores(programmingExerciseBeforeUpdate, updatedProgrammingExercise);
+        slideApi.ifPresent(api -> api.handleDueDateChange(programmingExerciseBeforeUpdate, updatedProgrammingExercise));
+
+        return ResponseEntity.ok(savedProgrammingExercise);
+    }
+
+    /**
+     * PUT /programming-exercises/timeline : Updates the timeline attributes of a given exercise
+     *
+     * @param updatedProgrammingExercise containing the changes that have to be saved
+     * @param notificationText           an optional text to notify the student group about the update on the programming exercise
+     * @return the ResponseEntity with status 200 (OK) with the updated ProgrammingExercise, or with status 403 (Forbidden)
+     *         if the user is not allowed to update the exercise or with 404 (Not Found) if the updated ProgrammingExercise couldn't be found in the database
+     */
+    @PutMapping("programming-exercises/timeline")
+    @EnforceAtLeastEditor
+    @FeatureToggle(Feature.ProgrammingExercises)
+    public ResponseEntity<ProgrammingExercise> updateProgrammingExerciseTimeline(@RequestBody ProgrammingExercise updatedProgrammingExercise,
+            @RequestParam(value = "notificationText", required = false) String notificationText) {
+        log.debug("REST request to update the timeline of ProgrammingExercise : {}", updatedProgrammingExercise);
+        var existingProgrammingExercise = programmingExerciseRepository.findByIdElseThrow(updatedProgrammingExercise.getId());
+        var user = userRepository.getUserWithGroupsAndAuthorities();
+        authCheckService.checkHasAtLeastRoleForExerciseElseThrow(Role.EDITOR, existingProgrammingExercise, user);
+        updatedProgrammingExercise = programmingExerciseCreationUpdateService.updateTimeline(updatedProgrammingExercise, notificationText);
+        exerciseService.logUpdate(updatedProgrammingExercise, updatedProgrammingExercise.getCourseViaExerciseGroupOrCourseMember(), user);
+        return ResponseEntity.ok().headers(HeaderUtil.createEntityUpdateAlert(applicationName, true, ENTITY_NAME, updatedProgrammingExercise.getTitle()))
+                .body(updatedProgrammingExercise);
+    }
+
+    /**
+     * PATCH /programming-exercises-problem: Updates the problem statement of the exercise.
+     *
+     * @param exerciseId              The ID of the exercise for which to change the problem statement
+     * @param updatedProblemStatement The new problemStatement
+     * @param notificationText        to notify the student group about the updated problemStatement on the programming exercise
+     * @return the ResponseEntity with status 200 (OK) and with body the updated problemStatement, with status 404 if the programmingExercise could not be found, or with 403 if the
+     *         user does not have permissions to access the programming exercise.
+     */
+    @PatchMapping("programming-exercises/{exerciseId}/problem-statement")
+    @EnforceAtLeastEditor
+    public ResponseEntity<ProgrammingExercise> updateProblemStatement(@PathVariable long exerciseId, @RequestBody String updatedProblemStatement,
+            @RequestParam(value = "notificationText", required = false) String notificationText) {
+        log.debug("REST request to update ProgrammingExercise with new problem statement: {}", updatedProblemStatement);
+        var programmingExercise = programmingExerciseRepository.findWithTemplateAndSolutionParticipationTeamAssignmentConfigCategoriesById(exerciseId)
+                .orElseThrow(() -> new EntityNotFoundException("Programming Exercise", exerciseId));
+        var user = userRepository.getUserWithGroupsAndAuthorities();
+        authCheckService.checkHasAtLeastRoleForExerciseElseThrow(Role.EDITOR, programmingExercise, user);
+        var updatedProgrammingExercise = programmingExerciseCreationUpdateService.updateProblemStatement(programmingExercise, updatedProblemStatement, notificationText);
+        exerciseService.logUpdate(updatedProgrammingExercise, updatedProgrammingExercise.getCourseViaExerciseGroupOrCourseMember(), user);
+        // we saved a problem statement with test ids instead of test names. For easier editing we send a problem statement with test names to the client:
+        programmingExerciseTaskService.replaceTestIdsWithNames(updatedProgrammingExercise);
+        return ResponseEntity.ok().headers(HeaderUtil.createEntityUpdateAlert(applicationName, true, ENTITY_NAME, updatedProgrammingExercise.getTitle()))
+                .body(updatedProgrammingExercise);
+    }
+
+    /**
+     * PUT /programming-exercises/{exerciseId}/generate-tests : Makes a call to StructureOracleGenerator to generate the structure oracle aka the test.json file
+     *
+     * @param exerciseId The ID of the programming exercise for which the structure oracle should get generated
+     * @return The ResponseEntity with status 201 (Created) or with status 400 (Bad Request) if the parameters are invalid
+     */
+    @PutMapping(value = "programming-exercises/{exerciseId}/generate-tests", produces = MediaType.TEXT_PLAIN_VALUE)
+    @EnforceAtLeastEditor
+    @FeatureToggle(Feature.ProgrammingExercises)
+    public ResponseEntity<String> generateStructureOracleForExercise(@PathVariable long exerciseId) {
+        log.debug("REST request to generate the structure oracle for ProgrammingExercise with id: {}", exerciseId);
+        var programmingExercise = programmingExerciseRepository.findByIdWithTemplateAndSolutionParticipationTeamAssignmentConfigCategoriesAndBuildConfigElseThrow(exerciseId);
+        User user = userRepository.getUserWithGroupsAndAuthorities();
+        authCheckService.checkHasAtLeastRoleForExerciseElseThrow(Role.EDITOR, programmingExercise, user);
+        if (programmingExercise.getPackageName() == null || programmingExercise.getPackageName().length() < 3) {
+            return ResponseEntity.badRequest().headers(HeaderUtil.createAlert(applicationName,
+                    "This is a linked exercise and generating the structure oracle for this exercise is not possible.", "couldNotGenerateStructureOracle")).body(null);
+        }
+
+        var solutionRepoUri = programmingExercise.getVcsSolutionRepositoryUri();
+        var exerciseRepoUri = programmingExercise.getVcsTemplateRepositoryUri();
+        var testRepoUri = programmingExercise.getVcsTestRepositoryUri();
+
+        try {
+            String testsPath = Path.of("test", programmingExercise.getPackageFolderName()).toString();
+            // Atm we only have one folder that can have structural tests, but this could change.
+            testsPath = programmingExercise.getBuildConfig().hasSequentialTestRuns() ? Path.of("structural", testsPath).toString() : testsPath;
+            boolean didGenerateOracle = programmingExerciseCreationUpdateService.generateStructureOracleFile(solutionRepoUri, exerciseRepoUri, testRepoUri, testsPath, user);
+
+            if (didGenerateOracle) {
+                HttpHeaders responseHeaders = new HttpHeaders();
+                responseHeaders.setContentType(MediaType.TEXT_PLAIN);
+                return new ResponseEntity<>("Successfully generated the structure oracle for the exercise " + programmingExercise.getProjectName(), responseHeaders, HttpStatus.OK);
+            }
+            else {
+                return ResponseEntity.badRequest().headers(
+                        HeaderUtil.createAlert(applicationName, "Did not update the oracle because there have not been any changes to it.", "didNotGenerateStructureOracle"))
+                        .body(null);
+            }
+        }
+        catch (Exception e) {
+            return ResponseEntity.badRequest()
+                    .headers(HeaderUtil.createAlert(applicationName,
+                            "An error occurred while generating the structure oracle for the exercise " + programmingExercise.getProjectName() + ": " + e,
+                            "errorStructureOracleGeneration"))
+                    .body(null);
+        }
+    }
+
+    /**
+     * PUT /programming-exercises/{exerciseId}/re-evaluate : Re-evaluates and updates an existing ProgrammingExercise.
+     *
+     * @param exerciseId                                  of the exercise
+     * @param programmingExercise                         the ProgrammingExercise to re-evaluate and update
+     * @param deleteFeedbackAfterGradingInstructionUpdate boolean flag that indicates whether the associated feedback should be deleted or not
+     * @return the ResponseEntity with status 200 (OK) and with body the updated ProgrammingExercise, or with status 400 (Bad Request) if the ProgrammingExercise is not valid,
+     *         or with status 409 (Conflict) if given exerciseId is not same as in the object of the request body, or with status 500 (Internal Server Error) if the
+     *         ProgrammingExercise
+     *         couldn't be updated
+     */
+    @PutMapping("programming-exercises/{exerciseId}/re-evaluate")
+    @EnforceAtLeastEditor
+    @FeatureToggle(Feature.ProgrammingExercises)
+    public ResponseEntity<ProgrammingExercise> reEvaluateAndUpdateProgrammingExercise(@PathVariable long exerciseId, @RequestBody ProgrammingExercise programmingExercise,
+            @RequestParam(value = "deleteFeedback", required = false) Boolean deleteFeedbackAfterGradingInstructionUpdate) throws JsonProcessingException {
+        log.debug("REST request to re-evaluate ProgrammingExercise : {}", programmingExercise);
+        // check that the exercise exists for given id
+        programmingExerciseRepository.findByIdElseThrow(exerciseId);
+        authCheckService.checkGivenExerciseIdSameForExerciseInRequestBodyElseThrow(exerciseId, programmingExercise);
+
+        // fetch course from database to make sure client didn't change groups
+        var user = userRepository.getUserWithGroupsAndAuthorities();
+        Course course = courseService.retrieveCourseOverExerciseGroupOrCourseId(programmingExercise);
+        authCheckService.checkHasAtLeastRoleInCourseElseThrow(Role.EDITOR, course, user);
+
+        exerciseService.reEvaluateExercise(programmingExercise, deleteFeedbackAfterGradingInstructionUpdate);
+        return updateProgrammingExercise(programmingExercise, null);
+    }
+
+}
